requirements.txt
torch
torchvision
numpy
pillow
matplotlib
scikit-learn
pyyaml
tqdm
